adapter: change logic around interpeting ATIM and ATIM_NOW (and MTIM equiv) to be consistient with upstream
in wasi-common preview1, ATIM and ATIM_NOW (and MTIM and MTIM now) were mutually exclusive and would result in an INVAL error, whereas in the adapter previously, ATIM_NOW implied ATIM, but would silently do nothing if ATIM was not set.

I decided to be consistient with the upstream behavior here because it is pretty arbitrary and I don't think there's a good reason to break compatibility.

This fixes the `path_filestat` test.